monitor/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
django-snippets==0.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
MarkupSafe==1.0
mysqlclient==1.3.13
pycparser==2.18
pyinotify==0.9.6
PyJWT==1.6.4
python-nmap==0.6.1
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
uWSGI==2.0.17.1
yagmail==0.11.214
